SKALE-4085 Start tests updating
