+++ proxy/test/MessageProxy.ts
@@ -31,8 +31,6 @@ import {
     DepositBoxEthInstance,
     ContractManagerInstance,
     LinkerInstance,
-    LockAndDataForSchainContract,
-    LockAndDataForSchainInstance,
     MessageProxyForMainnetInstance,
     MessageProxyForSchainContract,
     MessageProxyForSchainInstance,
@@ -58,13 +56,11 @@ import { rechargeSchainWallet } from "./utils/skale-manager-utils/wallets";
 
 const MessageProxyForSchain: MessageProxyForSchainContract = artifacts.require("./MessageProxyForSchain");
 const TokenManager: TokenManagerContract = artifacts.require("./TokenManager");
-const LockAndDataForSchain: LockAndDataForSchainContract = artifacts.require("./LockAndDataForSchain");
 const MessagesTester: MessagesTesterContract = artifacts.require("./MessagesTester");
 const SkaleFeaturesMock: SkaleFeaturesMockContract = artifacts.require("./SkaleFeaturesMock");
 
 contract("MessageProxy", ([deployer, user, client, customer]) => {
     let messageProxyForSchain: MessageProxyForSchainInstance;
-    let lockAndDataForSchain: LockAndDataForSchainInstance;
 
     let depositBox: DepositBoxEthInstance;
     let contractManager: ContractManagerInstance;
@@ -402,8 +398,6 @@ contract("MessageProxy", ([deployer, user, client, customer]) => {
 
         beforeEach(async () => {
             messageProxyForSchain = await MessageProxyForSchain.new("MyChain", {from: deployer});
-            lockAndDataForSchain = await LockAndDataForSchain.new({from: deployer});
-            await lockAndDataForSchain.setContract("MessageProxy", messageProxyForSchain.address, {from: deployer});
         });
 
         it("should detect registration state by `isConnectedChain` function", async () => {
@@ -567,7 +561,6 @@ contract("MessageProxy", ([deployer, user, client, customer]) => {
 
         it("should get outgoing messages counter", async () => {
             const chainID = randomString(10);
-            const contractAddress = lockAndDataForSchain.address;
             const amount = 5;
             const addressTo = client;
             const bytesData = await messages.encodeTransferEthMessage(addressTo, amount);
@@ -583,7 +576,7 @@ contract("MessageProxy", ([deployer, user, client, customer]) => {
             outgoingMessagesCounter0.should.be.deep.equal(new BigNumber(0));
 
             await messageProxyForSchain
-            .postOutgoingMessage(chainID, contractAddress, bytesData, {from: deployer});
+                .postOutgoingMessage(chainID, depositBox.address, bytesData, {from: deployer});
 
             const outgoingMessagesCounter = new BigNumber(
                 await messageProxyForSchain.getOutgoingMessagesCounter(chainID));
